+++ solver/tests/sat_solving_test.rs
@@ -9,7 +9,7 @@ const MAX_NUM_VARIABLES: usize = 3;
 const MAX_NUM_LITERALS: usize = 3;
 const MAX_NUM_CLAUSES: usize = 5;
 
-fn execute_solvers(formula: Vec<CNFClause>, num_variables: u32) -> (Option<Assignment>, Option<Assignment>) {
+fn execute_solvers(formula: Vec<CNFClause>, num_variables: usize) -> (Option<Assignment>, Option<Assignment>) {
     let testing_solver = CadicalSolver; // TODO: Replace testing_solver with custom solver
     let reference_solver = CadicalSolver;
 
@@ -21,7 +21,7 @@ fn execute_solvers(formula: Vec<CNFClause>, num_variables: u32) -> (Option<Assig
     
 proptest! {
     #[test]
-    fn only_positive_unit_clauses(num_variables in 1..=MAX_NUM_VARIABLES as u32) {
+    fn only_positive_unit_clauses(num_variables in 1..=MAX_NUM_VARIABLES ) {
         let formula = (1..=num_variables)
             .map(|variable| CNFClause{vars: vec![CNFVar::Pos(variable)]})
             .collect();
@@ -32,7 +32,7 @@ proptest! {
     }
     
     #[test]
-    fn only_negative_unit_clauses(num_variables in 1..=MAX_NUM_VARIABLES as u32) {
+    fn only_negative_unit_clauses(num_variables in 1..=MAX_NUM_VARIABLES ) {
         let formula = (1..=num_variables)
             .map(|variable| CNFClause{vars: vec![CNFVar::Neg(variable)]})
             .collect();
@@ -44,7 +44,7 @@ proptest! {
 
     #[test]
     fn only_unit_clauses(signs in vec(weighted(0.5), 1..=MAX_NUM_VARIABLES)) {
-        let num_variables = signs.len() as u32;
+        let num_variables = signs.len() ;
         let formula = signs.iter()
             .enumerate()
             .map(|(variable, sign)| {
@@ -52,7 +52,7 @@ proptest! {
                     true => CNFVar::Pos,
                     false => CNFVar::Neg,
                 };
-                CNFClause{vars: vec![literal((variable as u32)+1)]}
+                CNFClause{vars: vec![literal((variable )+1)]}
             }).collect();
 
         let (custom, reference) = execute_solvers(formula, num_variables);
@@ -63,7 +63,7 @@ proptest! {
     fn arbitrary_cnf_formula(clauses in vec(vec((1..=MAX_NUM_VARIABLES, weighted(0.5)), 1..=MAX_NUM_LITERALS), 1..=MAX_NUM_CLAUSES)) {
         let num_variables = clauses.iter()
             .max_by_key(|clause| clause.len())
-            .expect("Illegally, there were zero clauses!").len() as u32;
+            .expect("Illegally, there were zero clauses!").len() ;
 
         let formula = clauses.iter()
             .map(|clause| 
@@ -74,7 +74,7 @@ proptest! {
                             true => CNFVar::Pos,
                             false => CNFVar::Neg,
                         };
-                        literal(*variable as u32)
+                        literal(*variable )
                     }).collect()
                 }
             ).collect();
